requirements.txt
autopep8==1.3.3
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
enum34==1.1.6
fbchat==1.1.1
idna==2.6
pycodestyle==2.3.1
pycrypto==2.6.1
PyYAML==5.1
urllib3==1.24.2
requests==2.20.0
discord